tools/DeveloperTools/cloudflaretrace/api.py
import requests
from typing import Dict

def cloudflaretrace(endpoint: str = "https://one.one.one.one/cdn-cgi/trace", toolbench_rapidapi_key: str = '088440d910mshef857391f2fc461p17ae9ejsnaebc918926ff') -> Dict:
    """
    Retrieve trace information from Cloudflare including IP Address, Timestamp, User Agent, Country Code, and more.
    
    Parameters:
    endpoint [Optional]: str - The endpoint URL to query Cloudflare trace information.
    
    Returns:
    A dictionary with trace information.
    
    Note: This function accesses public endpoints and provides no additional authentication or use of the API key.
    """
    response = requests.get(endpoint)
    try:
        # Convert response text to dictionary
        data = {}
        for line in response.text.strip().split('\n'):
            key, value = line.split('=')
            data[key] = value
        return data
    except Exception as e:
        return {"error": str(e), "response": response.text}